+++ NEWS.md
@@ -3,6 +3,8 @@
 - Exported functions `get_file_name()` and `insert_empty_rows()`.
 - Updated `gentlg_single()` to replace, in the first column of tables/listings headers, every leading whitespace with 90 twips (0.0625 inches) left-indentation RTF markup.
 - `gentlg()` argument `wcol` now allows for a list of column width vectors when `huxme` is a list of tables.
+- `gentlg()` new argument `footers_one_row` which allows to merge footers as a single table row (default = FALSE)
+
 
 # tidytlg 0.11.0